# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: University Course Performance Tracker & Grade Prediction API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: major
          schema: { type: string }
        - in: query
          name: year
          schema: { type: integer }
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /students/{id}/gpa:
    get:
      operationId: getStudentGpa
      tags: [students]
      summary: Get student GPA and grade summary
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Student GPA details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentGpa"
  /students/{id}/courses:
    get:
      operationId: getStudentCourses
      tags: [students]
      summary: Get courses a student is enrolled in
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Student's courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentCourseDetail"

  # ─── Courses ─────────────────────────────────────────────────────
...
  /grades:
    get:
      operationId: listGrades
      tags: [grades]
      summary: List grades
      parameters:
        - in: query
          name: student_id
          schema: { type: integer }
        - in: query
          name: course_id
          schema: { type: integer }
        - in: query
          name: assignment_id
          schema: { type: integer }
      responses:
        "200":
          description: Grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
    post:
      operationId: createGrade
      tags: [grades]
      summary: Submit a grade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
  /grades/{id}:
    put:
      operationId: updateGrade
      tags: [grades]
      summary: Update a grade
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
    delete:
      operationId: deleteGrade
      tags: [grades]
      summary: Delete a grade
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Analytics ───────────────────────────────────────────────────
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, email, student_id, year, major, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        student_id: { type: string }
        year: { type: integer }
        major: { type: string }
        gpa: { type: ["number", "null"] }
        created_at: { type: string }
    StudentInput:
      type: object
      required: [name, email, student_id, year, major]
      properties:
        name: { type: string }
        email: { type: string }
        student_id: { type: string }
        year: { type: integer }
        major: { type: string }
    StudentUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        year: { type: integer }
        major: { type: string }
    StudentGpa:
      type: object
      required: [student_id, gpa, total_courses, completed_courses, letter_grade]
      properties:
        student_id: { type: integer }
        gpa: { type: number }
        total_courses: { type: integer }
        completed_courses: { type: integer }
        letter_grade: { type: string }
        grade_points_breakdown:
          type: array
          items:
            $ref: "#/components/schemas/CourseGradePoint"
    CourseGradePoint:
      type: object
      required: [course_name, grade, grade_points, credits]
      properties:
        course_name: { type: string }
        grade: { type: string }
        grade_points: { type: number }
        credits: { type: integer }
    StudentCourseDetail:
      type: object
      required: [course_id, course_name, course_code, semester, current_grade, letter_grade, credits]
      properties:
        course_id: { type: integer }
        course_name: { type: string }
        course_code: { type: string }
        semester: { type: string }
        current_grade: { type: ["number", "null"] }
        letter_grade: { type: ["string", "null"] }
        credits: { type: integer }

    # ── Course ───────────────────────────────────────────────────────
...
    Grade:
      type: object
      required: [id, student_id, assignment_id, score, submitted_at]
      properties:
        id: { type: integer }
        student_id: { type: integer }
        assignment_id: { type: integer }
        score: { type: number }
        percentage: { type: ["number", "null"] }
        letter_grade: { type: ["string", "null"] }
        feedback: { type: ["string", "null"] }
        submitted_at: { type: string }
        student_name: { type: ["string", "null"] }
        assignment_name: { type: ["string", "null"] }
        course_id: { type: ["integer", "null"] }
    GradeInput:
      type: object
      required: [student_id, assignment_id, score]
      properties:
        student_id: { type: integer }
        assignment_id: { type: integer }
        score: { type: number }
        feedback: { type: string }
    GradeUpdate:
      type: object
      properties:
        score: { type: number }
        feedback: { type: string }

    # ── Analytics ─────────────────────────────────────────────────────
...