# lib/api-zod/src/generated/types/courseGradePoint.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * University Course Performance Tracker & Grade Prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface CourseGradePoint {
  course_name: string;
  grade: string;
  grade_points: number;
  credits: number;
}