# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * University Course Performance Tracker & Grade Prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  name: string;
  email: string;
  student_id: string;
  year: number;
  major: string;
}